+++ src/main/java/com/github/cc3002/finalreality/controller/phases/WaitingPhase.java
@@ -1,0 +1,12 @@
+package com.github.cc3002.finalreality.controller.phases;
+
+public class WaitingPhase extends AbstractPhase {
+
+  /**
+   * Start a new turn, changing to decision phase.
+   */
+  @Override
+  public void newTurn() {
+    changePhase(new DecisionPhase());
+  }
+}