+++ src/main/java/com/github/cc3002/finalreality/model/character/player/AbstractPlayerCharacter.java
@@ -1,5 +1,6 @@
 package com.github.cc3002.finalreality.model.character.player;
 
+import com.github.cc3002.finalreality.controller.Controller;
 import com.github.cc3002.finalreality.model.character.AbstractCharacter;
 import com.github.cc3002.finalreality.model.character.Enemy;
 import com.github.cc3002.finalreality.model.character.ICharacter;
@@ -77,6 +78,7 @@ public abstract class AbstractPlayerCharacter extends AbstractCharacter implemen
       int damage = weapon.getDamage();
       target.receiveDamage(damage);
     }
+    finishTurnListened.firePropertyChange("NextTurn",null,this);
   }
 
   @Override
@@ -87,5 +89,10 @@ public abstract class AbstractPlayerCharacter extends AbstractCharacter implemen
   @Override
   public int getMaxMana () {
     throw new AssertionError("No Mana");
+  }
+
+  @Override
+  public void doPhase (Controller controller) {
+    controller.doPlayerPhase();
   }
 }